requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
faster-whisper>=1.0.0
librosa>=0.10.0
numpy>=1.24.0
soundfile>=0.12.0
pydub>=0.25.0
praat-parselmouth>=0.4.3
google-generativeai>=0.5.0
supabase>=2.0.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
pyaudio>=0.2.14
